requirements/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
astmonkey==0.3.6
astroid==3.0.1
bcrypt==4.0.1
black==23.10.0
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
coverage==7.3.2
cryptography==41.0.5
dill==0.3.7
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==20.0.3
fastapi==0.104.0
filelock==3.12.4
flake8==6.1.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
identify==2.5.30
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
MutPy==0.6.1
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.9.9
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
pip==22.0.2
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
psycopg2==2.9.9
pyasn1==0.5.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydot==1.4.2
pyflakes==3.1.0
pylint==3.0.2
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
setuptools==59.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
termcolor==2.4.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.23.2
virtualenv==20.24.5
watchfiles==0.21.0
websockets==12.0
